requirements.txt
requests
openai
PyMuPDF
python-dotenv
